Rename stores to playback
